backend/requirements_backup.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.6.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
influxdb-client==1.48.0
influxdb3-python==0.12.0
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.0
multidict==6.4.4
numpy==2.2.5
oauthlib==3.3.1
pandas==2.2.3
pillow==11.3.0
propcache==0.3.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
reactivex==4.0.4
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
yarl==1.20.0
zope.interface==7.2
